inline inter_canister_queries_test
