+++ rs/tests/execution/BUILD.bazel
@@ -85,14 +85,24 @@ system_test(
 
 system_test(
     name = "inter_canister_queries_test",
+    srcs = glob(["inter_canister_queries_tests/*.rs"]),
+    crate_root = "inter_canister_queries_test.rs",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
     ],
     target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
     runtime_deps = GUESTOS_RUNTIME_DEPS,
-    deps = DEPENDENCIES + ["//rs/tests"],
+    deps = [
+        "//rs/tests/driver:ic-system-test-driver",
+        "//rs/tests/execution/common:execution_system_tests_common",
+        "//rs/types/types",
+        "//rs/universal_canister/lib",
+        "@crate_index//:anyhow",
+        "@crate_index//:assert_matches",
+        "@crate_index//:ic-agent",
+        "@crate_index//:wat",
+    ],
 )
 
 system_test(
