+++ rs/tests/execution/inter_canister_queries_tests/call_on_cleanup.rs
