+++ rs/tests/execution/inter_canister_queries_tests/inter_canister_queries.rs
@@ -1,5 +1,3 @@
-/* tag::catalog[]
-end::catalog[] */
 use ic_agent::agent::RejectCode;
 use ic_system_test_driver::driver::test_env::TestEnv;
 use ic_system_test_driver::driver::test_env_api::{GetFirstHealthyNodeSnapshot, HasPublicApiUrl};
