+++ rs/tests/execution/inter_canister_queries_tests/mod.rs
@@ -1,0 +1,2 @@
+pub mod call_on_cleanup;
+pub mod inter_canister_queries;
